+++ resource/houdini_path/python2.7libs/pythonrc.py
@@ -5,16 +5,15 @@ import hou
 import ftrack
 import tempfile
 
-import ftrack_connect.config
-from ftrack_connect.ui.widget.import_asset import FtrackImportAssetDialog
-from ftrack_connect.ui.widget.asset_manager import FtrackAssetManagerDialog
+import ftrack_connector_legacy.config
+from ftrack_connector_legacy.ui.widget.import_asset import FtrackImportAssetDialog
+from ftrack_connector_legacy.ui.widget.asset_manager import FtrackAssetManagerDialog
 
 from ftrack_connect_houdini.connector import Connector
 from ftrack_connect_houdini.ui.info import FtrackHoudiniInfoDialog
 from ftrack_connect_houdini.ui.publisher import PublishAssetDialog
 from ftrack_connect_houdini.ui.tasks import FtrackTasksDialog
 
-from ftrack_connect_houdini.usage import send_event
 
 try:
     ftrack.setup()
@@ -27,11 +26,6 @@ connector.registerAssets()
 currentEntity = ftrack.Task(
     os.getenv('FTRACK_TASKID',
               os.getenv('FTRACK_SHOTID')))
-
-send_event(
-    'USED-FTRACK-CONNECT-HOUDINI'
-)
-
 
 
 def setFrameRangeData():
@@ -150,6 +144,6 @@ def showDialog(name):
 
     return dialog
 
-ftrack_connect.config.configure_logging(
+ftrack_connector_legacy.config.configure_logging(
     'ftrack_connect_houdini', level='WARNING'
 )